+++ packages/postgres/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @meshql/core@0.6.0
+
 ## 0.5.1
 
 ### Patch Changes